ensure user selects an initial type (even following OAuth log-in outside of sign-up)
A user may log in via an OAuth provider whether the identity they
authenticate via the provider has yet been added to our system or not.

We suggest the following workflow to the OAuth user, though they might not
follow it:

* sign up via the sign-up page(s): we'll save your initial type
preference in the session, and later set it on your new account, once
you authenticate with the provider

* subsequently log in via the log-in page: you'll authenticate via the
provider and that's that

However, there are other possibilities.

Even though they went through the sign-up page, they might return to us
having authenticated an account we've already seen. In this case, we'll
save the new "social account" data, if necessary, but otherwise do
nothing, other than log the user in.

On the other hand, the user may return to us from an OAuth provider
having authenticated an identity we've never seen; but, they initiated
the OAuth process through the "log in" page. In this case, they haven't
told us what user type they'd like, and there's no such information
stored in their session.

Moreover, there's a certain fragility to storing user type preference in
the session. Though unlikely, it's *possible* to lose this information.
While the ideal is to remember this information and *not* to ask the
user their preference again, we want a surefire check and a resolution
workflow, in case the user did not indicate their type preference, or in
case this information was somehow lost.

This changeset is that check and resolution workflow:

* `User.initial_type` may now be `NULL`

* If the session doesn't indicate a user type preference, for a new user account generated from OAuth, we set `initial_type` to `NULL`

* General view middleware checks, (regardless of request path), whether
an authenticated user's `initial_type` is `NULL`; and, if so, redirects
them to a page where they can select it. (Thanks to this middleware, we
can be sure that the user will proceed no further before resolving the
situation.)

* The new request handler and template set, a variation on the
pre-existing type selection page, sets the `initial_type`, and redirects
the user on to wherever they were originally intending to go.