domain package refactoring into Namespace (initial work) (#13)
> and minor refactoring of user creation in preparation for oauth integration

The `domain` package is no small thing, so this is a partial, or initial change, rather than a complete one.

The main thing is that using a class declaration, and `staticmethod` after static method, to give a namespace to what are really just simple functions, is a complete Python anti-pattern.

This is in part why Python modules exist and work the way they do. Put your functions in a file, import that file as a module, and voila -- you have your well-namespaced functions. And, if we were starting fresh, I would likely leave it at that.

That said, there are some limitations / particularities to using modules as namespaces in that way. And so, in sympathy, and with the hope of making something more flexible and more precise, which achieves the same without the specific and (in Python) here unnecessary language feature of the static class method, I put together a basic namespacing library, and began to apply it to the user domain module.

The main goals:

* functions are just (decorated) functions
* (and anything can be set on the namespace, not just functions)
* the namespace reflects only what you've added to it (no module imports, internals, _etc._)
* namespaces can be linked for simplicity and -- you know -- namespacing