integration with django-allauth for OAuth sign-up
Beginning with Google & GitHub

OAuth providers don't have a stable concept of a "username" and so that
has to be supplied. allauth doesn't assume you're collecting email
address (or something ...?) and so by default it creates a bizarre username
for you. Previously when I've used it, I didn't bother with a username
field at all. But for this, I've inserted a signal receiver which sets a
reasonable username, from the email address, immediately after the user
returns from the OAuth provider, but before their record has been saved.

And because in the sign-up flow the user selects their account type
first, (and perhaps regardless), the nicest thing for the user, and
simply the most pragmatic thing for us, was to store this selection in
their session, if and only if they request to sign up via OAuth.
(I dislike this overall but it's generally how these sorts of redirects
through OAuth are handled.) In the same signal receiver, we can
thereby set their User.initial_type appropriately, (once they have a
User, and before that field value is needed).

The general post-save signal receiver can then appropriately ensure the
user's volunteer profile, once the user's record has been saved.