round out social (oauth) login experience
* do not require user to provide existing password when setting a new
password -- iff they never had a password to begin with (account derived
from social login)

* provide page thru which to manage (add/remove) connected social accounts

* handle case of log-in (authentication) via *new* social account which
provides a *known* email address. allauth refuses this, treating it as
essentially a hacking attempt, and attempts to ask the user to
_provide an alternate email_; but, this doesn't really make sense to us,
as this likely is the (as yet unauthorized) account owner, (though we
won't *presume* it). instead, we'll prompt them to authenticate via the
existing account which claims the matching email address, and attempt to
immediately redirect them back to the oauth provider's login screen
following authentication, but as part of the social account management
"connect" process.