social auth (#19)
### Overview

Implements user authentication, both before and after sign-up, via OAuth connection of "social" app providers – currently Github & Google.

Resolves #10.

Ties together #16 and #18, with added refactoring and refinement.

### Sign-up

OAuth providers don't have a stable concept of a "username," and so that
has to be supplied, (_e.g._ for Google); and, allauth doesn't assume you're collecting email
address, and so by default it creates a bizarre username
for you. Previously when I've used it, I didn't bother with a username
field at all. But for this, via an allauth adapter, we'll set a
reasonable username, from the email address, immediately after the user
returns from the OAuth provider, but before their record has been saved.

And because in the sign-up flow the user selects their account type
first, (and perhaps regardless), the nicest thing for the user, and
simply the most pragmatic thing for us, was to store this selection in
their session, if and only if they request to sign up via OAuth.
(I dislike this overall but it's generally how these sorts of redirects
through OAuth are handled.) In the allauth adapter, we can
thereby set their `User.initial_type` appropriately, (once they have a
User, and before that field value is needed).

### Log-in

A user may log in via an OAuth provider whether the identity they
authenticate via the provider has yet been added to our system or not.

We suggest the following workflow to the OAuth user, though they might not
follow it:

* sign up via the sign-up page(s): we'll save your initial type
preference in the session, and later set it on your new account, once
you authenticate with the provider

* subsequently log in via the log-in page: you'll authenticate via the
provider and that's that

However, there are other possibilities.

Even though they went through the sign-up page, they might return to us
having authenticated an account we've already seen. In this case, we'll
save the new "social account" data, if necessary, but otherwise do
nothing, other than log the user in.

On the other hand, the user may return to us from an OAuth provider
having authenticated an identity we've never seen; but, they initiated
the OAuth process through the "log in" page. In this case, they haven't
told us what user type they'd like, and there's no such information
stored in their session.

Moreover, there's a certain fragility to storing user type preference in
the session. Though unlikely, it's *possible* to lose this information.
While the ideal is to remember this information and *not* to ask the
user their preference again, we want a surefire check and a resolution
workflow, in case the user did not indicate their type preference, or in
case this information was somehow lost.

This change is that check and resolution workflow:

* `User.initial_type` may now be `NULL`

* If the session doesn't indicate a user type preference, for a new user account generated from OAuth, we set `initial_type` to `NULL`

* General view middleware checks, (regardless of request path), whether
an authenticated user's `initial_type` is `NULL`; and, if so, redirects
them to a page where they can select it. (Thanks to this middleware, we
can be sure that the user will proceed no further before resolving the
situation.)

* The new request handler and template set, a variation on the
pre-existing type selection page, sets the `initial_type`, and redirects
the user on to wherever they were originally intending to go.

### Etc.

* Set password flow for OAuth-provisioned user – do not require user to provide existing password when setting a new password, iff they never had a password to begin with (account derived from social login)

* Page thru which to manage (add/remove) connected social accounts

* Handle case of log-in (authentication) via *new* social account which provides a *known* email address.

    allauth refuses this, treating it as essentially a hacking attempt, and attempts to ask the user to _provide an alternate email_; but, this doesn't really make sense to us, as this likely is the (as yet unauthorized) account owner, (though we won't *presume* it).

    Instead, we'll prompt them to authenticate via the existing account which claims the matching email address, and attempt to immediately redirect them back to the oauth provider's login screen following authentication, but as part of the social account management "connect" process.

### Fixes

* flash message alerts on "anonymous" home page (site index) were cut off by logo
* add "btn" class to pre-signup type selection buttons to ensure proper UI
* HTML/UI of password reset page was broken

### Changesets

0491f28 tightened up flash messages template logic
7e42faa typo correction and trivial template refactoring
46bf75e round out social (oauth) login experience
08da63d tweaked documentation string
4e7061d add "btn" class to pre-signup type selection buttons to ensure proper UI
2e7fa22 fix styling of flash message alerts on "anonymous" home page (site index)
0c85f78 ensure volunteer profile upon post-signup type selection
28a3a70 set volunteering task preferences in session prior to oauth signup
55ecef9 use font awesome branding icon for github oauth provider buttons/links
e101847 updated migration following rebase against parallel work
b712725 integrate allauth via social account adapter (rather than via signal)
4a4fe73 handle post-socialauth processing of new user via allauth adapter rather than via signal
879324d added dispatch_uids to signal connections to prevent duplication
72e9296 moved socialprovider_urls to new subpackage socialauth
1c0c236 moved signal connection from signals module to domain package
9ddc629 comments
9d4b6c3 ensure user selects an initial type (even following OAuth log-in outside of sign-up)
2149a1c refactored user signup type selection template to allow for post-oauth overrides
d1c5878 protect against garbage preference keys
dad3a6a styled OAuth buttons with provider logos (GitHub & Google)
ccb9f9c added social auth (OAuth) links to login page
58a2343 integration with django-allauth for OAuth sign-up
195907f create new user volunteer profile via signal -- for easy integration with OAuth