fix bug (unhandled programming error) in volunteer profile editing view
this view's permission check initially only checked that the user ID in
the URL, (a vanity artifact), was the same as that of the requesting
user, (though the former could be set manually) -- *not* that the ID
of the requested *profile* was owned by the requesting user, (nor that
the params were consistent with each other).

(full permissability was checked in another method, which itself is only
invoked for GET requests.)

this was "fixed" by attempting to go through the actual profile object
(to compare its owner with the requesting user, for identity) -- which
fails, because this permission check is run right *before* the profile
object is retrieved from the database.

THIS fix ensures that:

1) the view loads
2) the view URL parameters must be consistent (or else 404)
3) the profile must be owned by the requesting user (or else 403)