suppress overly verbose deployment output
when the deploy command is delegated to (from build), its by-default
suppression of command output is lost. moreover, its deployment-specific
output is way too verbose -- and, the command already parses it to
re-present only what's useful.

the deploy command now detects its delegation and suppresses this
verbose output in that case.

a call to describe-services is further replaced by its factored version.