prevent static file over-caching (#45)
* gzip static assets and set header/metadata "no-cache"

the latter appears necessary to force user agents to check cache
validity with server (S3) -- (which uses E-Tags)

* whitespace tweaks to "touch" css files

associated changes to static file-handling won't propagate until static
files are modified. this should be fine; though, this means the
liability of actual changes being missed -- due to user agent caching --
remains, in the meanwhile. as such, these high-change files are updated,
to eagerly propagate the change to caching headers, and mitigate this
risk.