restrict build-deploy to (environment-specific) deployment-compatible tags
The management command already restricts itself to semantic versioning
tags; however, it returns an error if asked to build/deploy an
inappropriate tag for the target environment -- which would fail the
workflow run (and create a distraction).

Github Actions don't (yet) support sophisticated tag filters on Release-
triggered workflows, nor does it offer built-in regular expression
functions.

Nor does it appear that a step can request that a job be "skipped."

Rather, compatibility is determined by a (Python) shell step, and
subsequent steps conditioned on this first step's output.