refactor build-deploy workflow to entirely skip non-targets
previous set-up ran a (matrix) job for each target, checked whether to
continue against the target, and then conditionally skipped the rest of
the job, (because Github actions don't currently allow for sophisticated
tag filtering when the action is triggered by a release).

this merely cleans this up slightly, by moving the check to its own job,
which outputs the list of targets compatible with the published release
tag. this simplifies the actual build-deploy job, and should make the
workflow reporting in Github more obviously reflect what was actually
done, (e.g. for which targets the build-deploy job actually ran and
succeeded).