clean-up of marketplace sub-domains & rules
partly addresses confusion over "DSSG Staff" vs. site "is_staff" and
"is_superuser" -- but only in that logic now attempts to consider all of
the above. as a result, the currently-unused site "Admin" link & page
(as opposed to the Django admin site) are now visible to natural
staff/superusers (regular users not initially created as DSSG staff).
(at least, this means that it is now obvious to those with admin
privileges that they have them, as the Admin link is visible.)

also begins to address technical confusion built into this codebase over
the `rules` library and how best to integrate it (has_perm vs.
test_rule, etc.)

anticipates #66