allow site admins to force completion of in-progress projects
* the "finish project" link/button is made visible to site admins on the
info pages of in-progress projects -- (projects which are not yet in the
QA stage and therefore not otherwise available to be marked complete).
** icon and label made to reflect that this action skips a step

* the "finish project" confirmation page is made to reflect when a
project is still in progress to warn unwitting admins

* the associated view/controller is updated to permit this action by
site admins
** (in fact, this view previously checked only for the user permission
-- not for project status -- and as such this was always available, and
to everyone with the project permission, who might have thought to try
this URL path. the view now further confirms that either the project is
in the QA state, or that both the project is in-progress and the request
is made by a site admin.)

resolves #66